Reset the project to start from the scratch
