+++ app-example/(tabs)/index.tsx
@@ -1,0 +1,29 @@
+import { View, StyleSheet, Text } from 'react-native';
+
+export default function HomeScreen() {
+  return (
+    <View style={styles.titleContainer}>
+      <Text >Welcome to React Native!</Text> 
+    </View>
+       
+  );
+}
+
+const styles = StyleSheet.create({
+  titleContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    gap: 8,
+  },
+  stepContainer: {
+    gap: 8,
+    marginBottom: 8,
+  },
+  reactLogo: {
+    height: 178,
+    width: 290,
+    bottom: 0,
+    left: 0,
+    position: 'absolute',
+  },
+});
